# web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Koritsu API",
  "description": "Two entrances, one set of handlers: `/api` for the site (cookie session) and `/api/v1` for external clients (bearer token). Only `/api/v1` is versioned and documented.",
  "version": "2.0.0a5.4"
 },
 "paths": {
...
  "/api/uml/themes": {
   "get": {
    "tags": [
     "uml"
    ],
    "summary": "Palettes of the UML diagram builder",
    "description": "Every palette the builder accepts. The name goes into the theme field of the building routes. The css palette paints with CSS variables, so one diagram fits both a dark page and a light report.",
    "operationId": "uml_themes",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "type": "string"
         },
         "type": "array",
         "title": "Response Uml Themes"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/uml/classes/preview": {
   "post": {
    "tags": [
     "uml"
    ],
    "summary": "Draw a class diagram without saving it",
    "description": "Draws a UML class diagram from sources in the body and returns the drawio XML without touching the volume. Any signed-in user, at most 30 requests per minute. 400 invalid_source, 400 unknown_lang, 400 unknown_theme, 413 source_too_large, 422 diagram_failed, 429 rate_limited.",
    "operationId": "uml_classes_preview",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UmlPreviewIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Uml Classes Preview"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/uml/objects/preview": {
   "post": {
    "tags": [
     "uml"
    ],
    "summary": "Draw an object diagram without saving it",
    "description": "Draws a UML object diagram from sources in the body and returns the drawio XML without touching the volume. The first source is the entry point, the rest are neighbouring files. Any signed-in user, at most 30 requests per minute. 400 invalid_source, 400 unknown_lang, 400 unknown_theme, 413 source_too_large, 422 diagram_failed, 429 rate_limited.",
    "operationId": "uml_objects_preview",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UmlPreviewIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Uml Objects Preview"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{project_id}/uml/classes": {
   "post": {
    "tags": [
     "uml"
    ],
    "summary": "Build a class diagram and keep it in the project",
    "description": "Builds a UML class diagram from text materials of the project or from sources in the body, stores the drawio XML as a project artifact and writes the diagram down in the run journal of the project. There is no separate save step: what was built is kept, together with the sources and the palette it was built with. The answer carries the XML, the artifact id and the run_id of the journal entry: delete that entry to delete the diagram. Editor role. 400 invalid_source, 400 unknown_lang, 400 unknown_theme, 403 forbidden, 404 not_found, 413 source_too_large, 422 diagram_failed.",
    "operationId": "uml_classes",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UmlIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DiagramBuiltOut"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{project_id}/uml/objects": {
   "post": {
    "tags": [
     "uml"
    ],
    "summary": "Build an object diagram and keep it in the project",
    "description": "Builds a UML object diagram from text materials of the project or from sources in the body, stores the drawio XML as a project artifact and writes the diagram down in the run journal of the project. The first source is the entry point, the rest are neighbouring files. The answer carries the XML, the artifact id, the run_id of the journal entry and the notes of the tracer. Editor role. 400 invalid_source, 400 unknown_lang, 400 unknown_theme, 403 forbidden, 404 not_found, 413 source_too_large, 422 diagram_failed.",
    "operationId": "uml_objects",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UmlIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DiagramBuiltOut"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{project_id}/uml/{run_id}": {
   "put": {
    "tags": [
     "uml"
    ],
    "summary": "Rebuild a UML diagram that is already in the project",
    "description": "Builds the diagram again, from new sources or with a new palette, and replaces what the journal entry points at. The entry itself stays: its name and its number do not change, because this is the same diagram drawn again, not a second one. The kind of diagram is the one it was built as. Editor role. 400 invalid_id, 400 invalid_source, 400 unknown_lang, 400 unknown_theme, 403 forbidden, 404 not_found, 413 source_too_large, 422 diagram_failed.",
    "operationId": "uml_rebuild",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     },
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UmlIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DiagramBuiltOut"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "uml"
    ],
    "summary": "One kept UML diagram, with its sources and its XML",
    "description": "The diagram as it was built: the drawio XML, the sources in the order they were given and the settings. This is what opens the diagram back up for editing. Viewer role. 400 invalid_id, 404 not_found.",
    "operationId": "uml_one",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     },
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DiagramFullOut"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{project_id}/uml": {
   "get": {
    "tags": [
     "uml"
    ],
    "summary": "UML diagrams kept in this project",
    "description": "Every UML diagram built in this project, newest first: what it is called, when it was built, which kind it is and which artifact holds its XML. Flowcharts are not in this list: they have their own. Viewer role. 400 invalid_id, 404 not_found.",
    "operationId": "uml_list",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/DiagramOut"
         },
         "title": "Response Uml List"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/uml/themes": {
   "get": {
    "tags": [
     "v1",
     "uml"
    ],
    "summary": "Palettes of the UML diagram builder",
    "description": "Every palette the builder accepts. The name goes into the theme field of the building routes. The css palette paints with CSS variables, so one diagram fits both a dark page and a light report.",
    "operationId": "uml_themes_v1",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "type": "string"
         },
         "type": "array",
         "title": "Response Uml Themes V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/uml/classes/preview": {
   "post": {
    "tags": [
     "v1",
     "uml"
    ],
    "summary": "Draw a class diagram without saving it",
    "description": "Draws a UML class diagram from sources in the body and returns the drawio XML without touching the volume. Any signed-in user, at most 30 requests per minute. 400 invalid_source, 400 unknown_lang, 400 unknown_theme, 413 source_too_large, 422 diagram_failed, 429 rate_limited.",
    "operationId": "uml_classes_preview_v1",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UmlPreviewIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Uml Classes Preview V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/uml/objects/preview": {
   "post": {
    "tags": [
     "v1",
     "uml"
    ],
    "summary": "Draw an object diagram without saving it",
    "description": "Draws a UML object diagram from sources in the body and returns the drawio XML without touching the volume. The first source is the entry point, the rest are neighbouring files. Any signed-in user, at most 30 requests per minute. 400 invalid_source, 400 unknown_lang, 400 unknown_theme, 413 source_too_large, 422 diagram_failed, 429 rate_limited.",
    "operationId": "uml_objects_preview_v1",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UmlPreviewIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Uml Objects Preview V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}/uml/classes": {
   "post": {
    "tags": [
     "v1",
     "uml"
    ],
    "summary": "Build a class diagram and keep it in the project",
    "description": "Builds a UML class diagram from text materials of the project or from sources in the body, stores the drawio XML as a project artifact and writes the diagram down in the run journal of the project. There is no separate save step: what was built is kept, together with the sources and the palette it was built with. The answer carries the XML, the artifact id and the run_id of the journal entry: delete that entry to delete the diagram. Editor role. 400 invalid_source, 400 unknown_lang, 400 unknown_theme, 403 forbidden, 404 not_found, 413 source_too_large, 422 diagram_failed.",
    "operationId": "uml_classes_v1",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UmlIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DiagramBuiltOut"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}/uml/objects": {
   "post": {
    "tags": [
     "v1",
     "uml"
    ],
    "summary": "Build an object diagram and keep it in the project",
    "description": "Builds a UML object diagram from text materials of the project or from sources in the body, stores the drawio XML as a project artifact and writes the diagram down in the run journal of the project. The first source is the entry point, the rest are neighbouring files. The answer carries the XML, the artifact id, the run_id of the journal entry and the notes of the tracer. Editor role. 400 invalid_source, 400 unknown_lang, 400 unknown_theme, 403 forbidden, 404 not_found, 413 source_too_large, 422 diagram_failed.",
    "operationId": "uml_objects_v1",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UmlIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DiagramBuiltOut"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/projects/{project_id}/uml": {
   "get": {
    "tags": [
     "v1",
     "uml"
    ],
    "summary": "UML diagrams kept in this project",
    "description": "Every UML diagram built in this project, newest first: what it is called, when it was built, which kind it is and which artifact holds its XML. Flowcharts are not in this list: they have their own. Viewer role. 400 invalid_id, 404 not_found.",
    "operationId": "uml_list_v1",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/DiagramOut"
         },
         "title": "Response Uml List V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ErrorBody": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code",
      "description": "Stable machine-readable code, e.g. not_found"
     },
     "message": {
      "type": "string",
      "title": "Message",
      "description": "Human-readable text, English"
     },
     "where": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Where",
      "description": "Field or parameter the error is about, e.g. body.email"
     }
    },
    "type": "object",
    "required": [
     "code",
     "message"
    ],
    "title": "ErrorBody",
    "description": "Тело ошибки: что случилось и, если это применимо, где."
   },
   "ErrorOut": {
    "properties": {
     "error": {
      "$ref": "#/components/schemas/ErrorBody"
     }
    },
    "type": "object",
    "required": [
     "error"
    ],
    "title": "ErrorOut",
    "description": "Единственная форма отказа службы. Другой нет ни у одного маршрута."
   },
...
   "UmlPreviewIn": {
    "properties": {
     "sources": {
      "items": {
       "$ref": "#/components/schemas/UmlSourceIn"
      },
      "type": "array",
      "title": "Sources",
      "description": "Sources to draw"
     },
     "lang": {
      "type": "string",
      "title": "Lang",
      "description": "Source language: py, cs or cpp",
      "default": "py"
     },
     "theme": {
      "type": "string",
      "title": "Theme",
      "description": "Palette, see GET /api/uml/themes",
      "default": "dark"
     }
    },
    "type": "object",
    "required": [
     "sources"
    ],
    "title": "UmlPreviewIn",
    "description": "То же, но без проекта: материалов здесь нет, только исходники в теле."
   },
...
  }
 },
...
}